+++ app/Listeners/SendWelcomeEmail.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Listeners;
+
+use App\Mail\UserWelcome;
+use Illuminate\Auth\Events\Registered;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Support\Facades\Mail;
+
+class SendWelcomeEmail implements ShouldQueue
+{
+    use InteractsWithQueue;
+
+    /**
+     * Handle the event.
+     */
+    public function handle(Registered $event): void
+    {
+        Mail::to($event->user->email)
+            ->queue(new UserWelcome($event->user));
+    }
+}